+++ src/index.js
@@ -1,6 +1,6 @@
 /**
  * Tree Component for uxcore
- * @author 
+ * @author
  *
  * Copyright 2014-2015, Uxcore Team, Alinw.
  * All rights reserved.